Split world_init.rs into subdirectory modules (946 -> ~200-377 lines each)
Split the monolithic world_init.rs (946 lines) into a world_init/
directory with four focused modules:
- mod.rs (171 lines): init_world entry point, terrain generation, helpers
- roads.rs (227 lines): road helpers and Tel Aviv road network layout
- zoning.rs (214 lines): zone assignment logic for Tel Aviv neighborhoods
- spawning.rs (377 lines): building, utility, service, and citizen spawning

All pub items re-exported from mod.rs so external API is unchanged.
